clippy

+++ littlefs-rust-core/tests/test_seek.rs
@@ -697,7 +697,7 @@ fn test_seek_reentrant_write(#[case] count: u32) {
         if open_err.is_ok() {
             let sz = littlefs_rust_core::lfs_file_size(lfs, file);
             if sz != 0 {
-                assert_eq!(sz, (count * 11) as u32);
+                assert_eq!(sz, count * 11);
                 for _ in 0..count {
                     let n = littlefs_rust_core::lfs_file_read(lfs, file, &mut buf[..11])?;
                     if n != 11 {
@@ -727,10 +727,7 @@ fn test_seek_reentrant_write(#[case] count: u32) {
 
         littlefs_rust_core::lfs_file_open(lfs, file, path, LFS_O_RDWR)?;
 
-        assert_eq!(
-            littlefs_rust_core::lfs_file_size(lfs, file),
-            (count * 11) as u32
-        );
+        assert_eq!(littlefs_rust_core::lfs_file_size(lfs, file), count * 11);
 
         let mut off: u32 = 0;
         for _ in 0..count {
@@ -781,10 +778,7 @@ fn test_seek_reentrant_write(#[case] count: u32) {
         littlefs_rust_core::lfs_file_close(lfs, file)?;
 
         littlefs_rust_core::lfs_file_open(lfs, file, path, LFS_O_RDWR)?;
-        assert_eq!(
-            littlefs_rust_core::lfs_file_size(lfs, file),
-            (count * 11) as u32
-        );
+        assert_eq!(littlefs_rust_core::lfs_file_size(lfs, file), count * 11);
         for _ in 0..count {
             let n = littlefs_rust_core::lfs_file_read(lfs, file, &mut buf[..11])?;
             if n != 11 {
